Move enums out of nesting
